Address comments & move to helper
